started to cleanup a bit
