+++ NebulaPatcher/Patches/Transpilers/PlayerAction_Mine_Patch.cs
@@ -1,44 +1,14 @@
 ﻿using HarmonyLib;
 using LiteNetLib;
 using NebulaModel.Packets.Planet;
-using NebulaModel.Packets.Factory;
 using NebulaWorld;
 using System;
 using System.Collections.Generic;
 using System.Reflection;
 using System.Reflection.Emit;
 
-using UnityEngine;
-
 namespace NebulaPatcher.Patches.Dynamic
 {
-
-    [HarmonyPatch(typeof(PlanetFactory), "BuildFinally")]
-    class patch
-    {
-        public static bool Prefix(PlanetFactory __instance, Player player, int prebuildId)
-        {
-
-            if (prebuildId != 0)
-            {
-                PrebuildData data = __instance.prebuildPool[prebuildId];
-                if (data.id == prebuildId)
-                {
-                    OnEntityPlaced(data.protoId, data.pos, data.rot);
-                }
-            }
-
-            return true;
-
-        }
-
-        private static void OnEntityPlaced(short protoId, Vector3 pos, Quaternion rot)
-        {
-            var packet = new EntityPlaced(protoId, pos, rot);
-            LocalPlayer.SendPacket(packet, DeliveryMethod.ReliableUnordered);
-        }
-    }
-
     [HarmonyPatch(typeof(PlayerAction_Mine), "GameTick")]
     class PlayerAction_Mine_Transpiler
     {
